add test coverage
