+++ src/Frontend/Dfe.Complete/Pages/Projects/List/AllProjectsModel.cs
@@ -20,8 +20,8 @@ public abstract class AllProjectsModel(string currentNavigation) : BaseProjectsP
 
     public static string GetTrustProjectsUrl(ListTrustsWithProjectsResultModel trustModel)
     {
-        return trustModel.identifier.Contains("TR") ? 
-                                        string.Format(RouteConstants.TrustMATProjects, trustModel.identifier) 
-                                        : string.Format(RouteConstants.TrustProjects, trustModel.identifier);
+        return trustModel.identifier.Contains("TR") 
+            ? string.Format(RouteConstants.TrustMATProjects, trustModel.identifier)
+            : string.Format(RouteConstants.TrustProjects, trustModel.identifier);
     }
 }
